+++ src/orion/core/cli/db/dump.py
@@ -11,7 +11,7 @@ import logging
 
 from orion.core.cli import base as cli
 from orion.core.io import experiment_builder
-from orion.core.worker.storage_backup import dump_database
+from orion.storage.backup import dump_database
 from orion.storage.base import setup_storage
 
 logger = logging.getLogger(__name__)
